% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVarToFactor.R
\name{convertVarToFactor}
\alias{convertVarToFactor}
\title{Convert specified variable to factor,
with ordered specified by an associated numeric variable}
\usage{
convertVarToFactor(data, var, varNum)
}
\arguments{
\item{data}{data.frame with data}

\item{var}{character vector with variable(s)}

\item{varNum}{character vector with associated numeric variable(s)}
}
\value{
data.frame with specified variable converted to factor
}
\description{
Convert specified variable to factor,
with ordered specified by an associated numeric variable
}
\author{
Laure Cougnaud
}
