% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getGLPGColorPalette}
\alias{getGLPGColorPalette}
\title{Get color palette of specified length,
either from a vector of names for the palette, or
from a specified length.}
\usage{
getGLPGColorPalette(n = NULL, x = NULL, type = c("GLPG", "viridis"))
}
\arguments{
\item{n}{number of elements in the palette}

\item{x}{vector with elements used as names for the palette.
If factor, the levels are used, otherwise the unique elements of the vector.}

\item{type}{character with type of palette used:
'viridis' for color-blinded friendly palette or
'GLPG' for Galapagos color palette}
}
\value{
vector of viridis colors,
named with the elements in \code{x} if \code{x} is specified.
}
\description{
Get color palette of specified length,
either from a vector of names for the palette, or
from a specified length.
}
\examples{
getGLPGColorPalette(n = 10)
getGLPGColorPalette(x = paste('treatment', 1:4))
}
\author{
Laure Cougnaud
}
