% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{getGLPGLinetypePalette}
\alias{getGLPGLinetypePalette}
\title{Get linetype palette.}
\usage{
getGLPGLinetypePalette(n = NULL, x = NULL)
}
\arguments{
\item{n}{number of elements in the palette}

\item{x}{vector with elements used as names for the palette.
If factor, the levels are used, otherwise the unique elements of the vector.}
}
\value{
character vector values with linetype
}
\description{
Note that 7 unique symbols are available at maximum.
}
\author{
Laure Cougnaud
}
