% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitPrintListPlots.R
\name{knitPrintListObjects}
\alias{knitPrintListObjects}
\title{include list of objects with possibly different chunk options
in a knitr document.}
\usage{
knitPrintListObjects(xList, generalLabel = "objectsList",
  labels = paste0(generalLabel, seq_along(xList)), titles = NULL,
  titleLevel = 2, printObject = FALSE, ...)
}
\arguments{
\item{xList}{List of objects to print.}

\item{generalLabel}{general label for the chunks, used to build the \code{labels}.
The labels are constructed as '\code{generalLabel}[i]',
with i the list index.
Only use if \code{labels} is not specified.}

\item{labels}{vector with labels for each chunk}

\item{titles}{vector with section titles}

\item{titleLevel}{integer with level for section header}

\item{printObject}{Logical, if TRUE (FALSE by default),
each object within \code{xList} is explicitely printed 
with the \code{\link{print}} function.}

\item{...}{any chunk parameters, will be replicated if necessary
See \code{\link{knitr}[opts_chunk]} for further details on available options.}
}
\value{
no returned value, a text is printed with chunk content
}
\description{
include list of objects with possibly different chunk options
in a knitr document.
}
\author{
Laure Cougnaud
}
