% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitPrintListPlots.R
\name{knitPrintListPlots}
\alias{knitPrintListPlots}
\title{include list of plots with different figure height, size, label in a knitr document}
\usage{
knitPrintListPlots(plotsList, generalLabel = "plotsList",
  type = c("ggplot2", "plotly"), ...)
}
\arguments{
\item{plotsList}{list of plots, e.g. \code{\link{ggplot2}[ggplot]}}

\item{generalLabel}{general label for the chunks, used to build the \code{labels}.
The labels are constructed as '\code{generalLabel}[i]',
with i the list index.
Only use if \code{labels} is not specified.}

\item{type}{string with plot type: 'ggplot2', 'plotly'}

\item{...}{Additional parameters for the \code{\link{knitPrintListObjects}} function.}
}
\value{
no returned value, a text is printed with chunk content
}
\description{
Each plot will be included in a separated chunk.
Reason: fig.height and fig.width options are not vectorized.
}
\author{
Laure Cougnaud
}
